Refactor texture loading
Add a ResourceLoader class which is responsible for loading textures and
maintaining ownership of the GLMaterial objects.  This just consolidates
the logic in GLSurface into a separate class dedicated to texture
managemeent.

This also updates the Mesh to directly store GLMaterial pointers, rather
than storing an index into a material array.  This makes the code
simpler, and reduces the possibility for out of bounds array accesses.
(The old code had some bugs where it could access past the end of the
array if it failed to load materials.)