Avoid use of multicharacter literals in TriFile.cpp
The behavior of multicharacter literals is implementation-defined
according to the C++ standard, so there is no guarantees about how they
will behave.  Apparently Visual Studio parses them in host byte-order,
so they are little endian on x86.

This fixes the code to avoid using multicharacter literals, and confirms
that the first four bytes of a TriFile are "PIRT" or "\0IRT".  This will
always behave correctly whether or not it is compiled on a little-endian
or big-endian machine.