Avoid using variables named "small" on windows
Some of the Windows header files (rpcndr.h) define "small" as a macro.
(#define small char).  This breaks code that tries to use small as a
variable name if the windows headers have already been included.

This renames the variable to "little", just to avoid this conflict.
Otherwise SliderManager.h can break, depending simply on the include
order.